Authorization with Policy

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateTaskRequest;
 use App\Http\Resources\TaskResource;
 use App\Models\Task;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -16,6 +17,8 @@ class TaskController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Task::class);
+
         return TaskResource::collection(auth()->user()->tasks()->get());
     }
 
@@ -24,6 +27,10 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        if (request()->user()->cannot('create', Task::class)) {
+            abort(403, "This action is unauthorized");
+        }
+
         $task = $request->user()->tasks()->create($request->validated());
 
         return TaskResource::make($task);
@@ -34,6 +41,8 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        Gate::authorize('view', $task);
+
         return TaskResource::make($task);
     }
 
@@ -42,6 +51,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if ($request->user()->cannot('update', $task)) {
+            abort(403, "This action is unauthorized");
+        }
+
         $task->update($request->validated());
 
         return TaskResource::make($task);
@@ -52,6 +65,10 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
+        if (request()->user()->cannot('delete', $task)) {
+            abort(403, "This action is unauthorized");
+        }
+
         $task->delete();
 
         return response()->noContent();
